clients controllers worst with new database

+++ app/Http/Controllers/ClientsController.php
@@ -6,8 +6,10 @@ use Illuminate\Http\Request;
 
 use Auth;
 use JavaScript;
+use DB;
 
 use App\Client;
+use App\User;
 use App\Http\Requests;
 
 use App\Http\Requests\CreateClientRequest;
@@ -24,7 +26,7 @@ class ClientsController extends Controller
     {
         $this->middleware('auth');
     }
-    
+
     /**
      * Display a listing of the resource.
      *
@@ -32,10 +34,24 @@ class ClientsController extends Controller
      */
     public function index()
     {
+        $user = Auth::user();
+        if($user->cannot('index', Client::class))
+        {
+            // abort(403);
+            return 'you should not pass';
+        }
+
+        if($user->isAdministrator())
+        {
+            $clients = $user->userable()->clients()->get();
+        }else{
+            $clients = $user->userable()->admin()->clients()->get();
+        }
+
         JavaScript::put([
             'click_url' => url('clients').'/',
         ]);
-        $clients = Auth::user()->clients();
+
         return view('clients.index', compact('clients'));
     }
 
@@ -46,6 +62,13 @@ class ClientsController extends Controller
      */
     public function create()
     {
+        $user = Auth::user();
+        if($user->cannot('create', Client::class))
+        {
+            // abort(403);
+            return 'you should not pass';
+        }
+
         return view('clients.create');
     }
 
@@ -57,16 +80,44 @@ class ClientsController extends Controller
      */
     public function store(CreateClientRequest $request)
     {
+        $user = Auth::user();
+        if($user->cannot('create', Client::class))
+        {
+            // abort(403);
+            return 'you should not pass';
+        }
+
+        if($user->isAdministrator())
+        {
+            $admin = $user->userable();
+        }else{
+            $admin = $user->userable()->admin();
+        }
+
         $client = Client::create([
             'name' => $request->name,
             'last_name' => $request->last_name,
-            'email' => $request->email,
             'cellphone' => $request->cellphone,
             'type' => $request->type,
             'language' => $request->language,
+		    'email_preferences' => 4,
             'comments' => $request->comments,
-            'user_id' => Auth::user()->id,
+            'admin_id' => $admin->id,
         ]);
+        $user = User::create([
+            'email' => $request->email,
+            'password' => bcrypt(str_random(9)),
+            'userable_id' => $client->id,
+            'userable_type' => 'App\Client',
+            'remember_token' => str_random(10),
+            'api_token' => str_random(60),
+        ]);
+
+        // DB::table('client_service')->insert([
+        //         'client_id' => $client_id,
+        //         'service_id' => $service_id,
+        //     ]);
+
         $photo = true;
         if($request->photo){
             $photo = $client->addImageFromForm($request->file('photo'));
@@ -87,7 +138,20 @@ class ClientsController extends Controller
      */
     public function show($seq_id)
     {
-        $client = Auth::user()->clientsBySeqId($seq_id);
+        $user = Auth::user();
+        if($user->cannot('show', Client::class))
+        {
+            // abort(403);
+            return 'you should not pass';
+        }
+
+        if($user->isAdministrator())
+        {
+            $client = $user->userable()->clientsBySeqId($seq_id);
+        }else{
+            $client = $user->userable()->admin()->clientsBySeqId($seq_id);
+        }
+
         return view('clients.show',compact('client'));
     }
 
@@ -99,7 +163,20 @@ class ClientsController extends Controller
      */
     public function edit($seq_id)
     {
-        $client = Auth::user()->clientsBySeqId($seq_id);
+        $user = Auth::user();
+        if($user->cannot('edit', Client::class))
+        {
+            // abort(403);
+            return 'you should not pass';
+        }
+
+        if($user->isAdministrator())
+        {
+            $client = $user->userable()->clientsBySeqId($seq_id);
+        }else{
+            $client = $user->userable()->admin()->clientsBySeqId($seq_id);
+        }
+
         return view('clients.edit',compact('client'));
     }
 
@@ -112,11 +189,26 @@ class ClientsController extends Controller
      */
     public function update(CreateClientRequest $request, $seq_id)
     {
-        $client = Auth::user()->clientsBySeqId($seq_id);
+        $user = Auth::user();
+        if($user->cannot('edit', Client::class))
+        {
+            // abort(403);
+            return 'you should not pass';
+        }
+
+        if($user->isAdministrator())
+        {
+            $client = $user->userable()->clientsBySeqId($seq_id);
+            $user = $client->user();
+        }else{
+            $client = $user->userable()->admin()->clientsBySeqId($seq_id);
+            $user = $client->user();
+        }
+
+        $user->email = $request->email;
 
         $client->name = $request->name;
         $client->last_name = $request->last_name;
-        $client->email = $request->email;
         $client->cellphone = $request->cellphone;
         $client->type = $request->type;
         $client->language = $request->language;
@@ -128,7 +220,7 @@ class ClientsController extends Controller
             $photo = $client->addImageFromForm($request->file('photo'));
         }
 
-        if($client->save() && $photo){
+        if($user->save() && $client->save() && $photo){
             flash()->success('Updated', 'New client successfully updated.');
             return redirect('clients/'.$seq_id);
         }
@@ -145,7 +237,20 @@ class ClientsController extends Controller
      */
     public function destroy($seq_id)
     {
-        $client = Auth::user()->clientsBySeqId($seq_id);
+        $user = Auth::user();
+        if($user->cannot('destroy', Client::class))
+        {
+            // abort(403);
+            return 'you should not pass';
+        }
+
+        if($user->isAdministrator())
+        {
+            $client = $user->userable()->clientsBySeqId($seq_id);
+        }else{
+            $client = $user->userable()->admin()->clientsBySeqId($seq_id);
+        }
+
         if($client->delete()){
             flash()->success('Deleted', 'The client was successfuly deleted');
             return redirect('clients');
